+++ packages/crypto-utils/src/index.js
@@ -1,3 +1,4 @@
+export * from './key-utils';
 export * from './keypairs';
 export * from './types';
 export * from './utils';